Version 1
This code performs a meticulous statistical analysis of the data obtained from a scan conducted by the Carimagua Radar (Meta, Colombia). The nature of the variables is analyzed, with an emphasis on reflectivity and intensity, in order to subsequently use the Z-R relationship for precipitation estimation in this region of the Colombian territory.